T3-01: 通过 ui_common 统一 DEG_LOG 日志 UI 更新
