+++ ui_common.h
@@ -13,3 +13,7 @@ GtkWidget* create_bottom_controls(GtkWidgetHelper& helper);
 
 // 底部控制栏信号绑定
 void bind_bottom_signals(GtkWidgetHelper& helper, GtkWidget* bottomContainer);
+
+// 日志 UI 操作：在日志标签页追加一行文本
+// type 通常对应 common.h 中的 msg_type（I/W/E/OP/DE），message 为已格式化好的单行文本
+void append_log_to_ui(int type, const char* message);
